requirements.txt
pdfminer.six>=20221105
beautifulsoup4>=4.12.0 